tweaking api
challenge summary now shows difficulty and no longer the centroid which
made no sense. The challenges endpoint no longer falls back to a
default difficulty if none is given or set in the user preferences,
rather returns all difficulties and lets the client deal with it. Also,
repaired the contains functionality.